requirements.txt
streamlit==1.35.0
python-docx==1.1.0
pandas==2.2.2
scikit-learn==1.4.2
matplotlib==3.9.0
numpy>=1.26
